Remove unit test class HelmOperatorTestSuiteMixin
The purpose of this change is to remove the
HelmOperatorTestSuiteMixin in the unit tests. This reference is
part of Armada and crashes Zuul because it was removed
here: https://review.opendev.org/c/starlingx/config/+/869094

It was necessary to make adjustments to configuration files for unit
tests and pylint because zuul was failing.

Test Plan:
PASS build-pkgs -p python3-k8sapp-oran-o2
PASS build-pkgs -p stx-oran-o2-helm
PASS unit test is performed successfully

Story: 2010560
Task: 48751

Co-Authored-By: Igor Pires Soares <igor.piressoares@windriver.com>
Change-Id: Idb0963c60f8a8a5ec3acd139f60383a128dc43b7
Signed-off-by: David Barbosa Bastos <david.barbosabastos@windriver.com>